Build User Interface

+++ app/src/main/java/com/tripiana/cabifytest/domain/model/BulkDiscountModel.kt
@@ -2,5 +2,5 @@ package com.tripiana.cabifytest.domain.model
 
 data class BulkDiscountModel(
     val units: Int,
-    val price: Double
+    val price: Float
 )
